Restructure utility files
