+++ packages/marko-web-omeda-identity-x/omeda-data/index.js
@@ -1,0 +1,13 @@
+const getAnsweredQuestionMap = require('./get-answered-question-map');
+const getOmedaCustomerRecord = require('./get-omeda-customer-record');
+const getOmedaLinkedFields = require('./get-omeda-linked-fields');
+const setOmedaDemographics = require('./set-omeda-demographics');
+const setOmedaDeploymentTypes = require('./set-omeda-deployment-types');
+
+module.exports = {
+  getAnsweredQuestionMap,
+  getOmedaCustomerRecord,
+  getOmedaLinkedFields,
+  setOmedaDemographics,
+  setOmedaDeploymentTypes,
+};
